+++ GRDB/FTS/FTS5CustomTokenizer.swift
@@ -163,7 +163,7 @@
         ///     class MyTokenizer : FTS5CustomTokenizer { ... }
         ///     dbPool.add(tokenizer: MyTokenizer.self)
         public func add<Tokenizer: FTS5CustomTokenizer>(tokenizer: Tokenizer.Type) {
-            write { db in
+            unsafeWrite { db in // Actually safe since we do not modify the database
                 db.add(tokenizer: Tokenizer.self)
             }
         }