+++ Tests/Public/Core/DatabasePool/DatabasePoolConcurrencyTests.swift
@@ -9,6 +9,208 @@ import XCTest
 
 class DatabasePoolConcurrencyTests: GRDBTestCase {
     
+    func testDatabasePoolFundamental1() {
+        // Constraint: the sum of values, the balance, must remain zero.
+        // DatabasePool aims at providing this guarantee, that is to say:
+        // no reader should ever see a non-zero balance.
+        //
+        // This test shows that writes that are not wrapped in a transaction
+        // can not provide our guarantee.
+        //
+        // Reader                                   Writer
+        // BEGIN DEFERRED TRANSACTION
+        let s1 = DispatchSemaphore(value: 0)
+        //                                          INSERT INTO moves VALUES (1)
+        let s2 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves
+        let s3 = DispatchSemaphore(value: 0)
+        //                                          INSERT INTO moves VALUES (-1)
+        let s4 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves
+        // END
+        
+        assertNoError {
+            dbConfiguration.trace = { print($0) }
+            do {
+                let dbQueue = try makeDatabaseQueue(filename: "test.sqlite")
+                try dbQueue.inDatabase { db in
+                    let journalMode = try String.fetchOne(db, "PRAGMA journal_mode = wal")
+                    XCTAssertEqual(journalMode, "wal")
+                    try db.create(table: "moves") { $0.column("value", .integer) }
+                    try db.execute("INSERT INTO moves VALUES (0)")
+                }
+            }
+            let s0 = DispatchSemaphore(value: 0)
+            let block1 = { () in
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                s0.signal() // Avoid "database is locked" error: don't open the two databases at the same time
+                try! dbQueue.inDatabase { db in
+                    try db.execute("BEGIN DEFERRED TRANSACTION")
+                    s1.signal()
+                    _ = s2.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 1)  // Non-zero balance
+                    s3.signal()
+                    _ = s4.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 1)  // Non-zero balance
+                    try db.execute("END")
+                }
+            }
+            let block2 = { () in
+                _ = s0.wait(timeout: .distantFuture) // Avoid "database is locked" error: don't open the two databases at the same time
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                try! dbQueue.inDatabase { db in
+                    _ = s1.wait(timeout: .distantFuture)
+                    try db.execute("INSERT INTO moves VALUES (1)")
+                    s2.signal()
+                    _ = s3.wait(timeout: .distantFuture)
+                    try db.execute("INSERT INTO moves VALUES (-1)")
+                    s4.signal()
+                }
+            }
+            let blocks = [block1, block2]
+            DispatchQueue.concurrentPerform(iterations: blocks.count) { index in
+                blocks[index]()
+            }
+        }
+    }
+
+    func testDatabasePoolFundamental2() {
+        // Constraint: the sum of values, the balance, must remain zero.
+        // DatabasePool aims at providing this guarantee, that is to say:
+        // no reader should ever see a non-zero balance.
+        //
+        // This test shows that writes that happen in a transaction are not
+        // visible from the reader.
+        //
+        // Reader                                   Writer
+        //                                          BEGIN DEFERRED TRANSACTION
+        let s1 = DispatchSemaphore(value: 0)
+        // BEGIN DEFERRED TRANSACTION
+        let s2 = DispatchSemaphore(value: 0)
+        //                                          INSERT INTO moves VALUES (1)
+        let s3 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves
+        let s4 = DispatchSemaphore(value: 0)
+        //                                          INSERT INTO moves VALUES (-1)
+        let s5 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves
+        // END                                      END
+        
+        assertNoError {
+            do {
+                let dbQueue = try! makeDatabaseQueue(filename: "test.sqlite")
+                try dbQueue.inDatabase { db in
+                    let journalMode = try String.fetchOne(db, "PRAGMA journal_mode = wal")
+                    XCTAssertEqual(journalMode, "wal")
+                    try db.create(table: "moves") { $0.column("value", .integer) }
+                    try db.execute("INSERT INTO moves VALUES (0)")
+                }
+            }
+            let s0 = DispatchSemaphore(value: 0)
+            let block1 = { () in
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                s0.signal() // Avoid "database is locked" error: don't open the two databases at the same time
+                try! dbQueue.inDatabase { db in
+                    _ = s1.wait(timeout: .distantFuture)
+                    try db.execute("BEGIN DEFERRED TRANSACTION")
+                    s2.signal()
+                    _ = s3.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 0)  // Zero balance
+                    s4.signal()
+                    _ = s5.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 0)  // Zero balance
+                    try db.execute("END")
+                }
+            }
+            let block2 = { () in
+                _ = s0.wait(timeout: .distantFuture) // Avoid "database is locked" error: don't open the two databases at the same time
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                try! dbQueue.inDatabase { db in
+                    try db.execute("BEGIN DEFERRED TRANSACTION")
+                    s1.signal()
+                    _ = s2.wait(timeout: .distantFuture)
+                    try db.execute("INSERT INTO moves VALUES (1)")
+                    s3.signal()
+                    _ = s4.wait(timeout: .distantFuture)
+                    try db.execute("INSERT INTO moves VALUES (-1)")
+                    s5.signal()
+                    try db.execute("END")
+                }
+            }
+            let blocks = [block1, block2]
+            DispatchQueue.concurrentPerform(iterations: blocks.count) { index in
+                blocks[index]()
+            }
+        }
+    }
+    
+    func testDatabasePoolFundamental3() {
+        // Constraint: the sum of values, the balance, must remain zero.
+        // DatabasePool aims at providing this guarantee, that is to say:
+        // no reader should ever see a non-zero balance.
+        //
+        // This test shows that writes that happen in a transaction are not
+        // visible from the reader.
+        //
+        // Reader                                   Writer
+        // BEGIN DEFERRED TRANSACTION
+        let s1 = DispatchSemaphore(value: 0)
+        //                                          BEGIN DEFERRED TRANSACTION
+        //                                          INSERT INTO moves VALUES (1)
+        let s2 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves
+        let s3 = DispatchSemaphore(value: 0)
+        //                                          INSERT INTO moves VALUES (-1)
+        let s4 = DispatchSemaphore(value: 0)
+        // SELECT SUM(value) AS balance FROM moves  END
+        // END
+        
+        assertNoError {
+            do {
+                let dbQueue = try! makeDatabaseQueue(filename: "test.sqlite")
+                try dbQueue.inDatabase { db in
+                    let journalMode = try String.fetchOne(db, "PRAGMA journal_mode = wal")
+                    XCTAssertEqual(journalMode, "wal")
+                    try db.create(table: "moves") { $0.column("value", .integer) }
+                    try db.execute("INSERT INTO moves VALUES (0)")
+                }
+            }
+            let s0 = DispatchSemaphore(value: 0)
+            let block1 = { () in
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                s0.signal() // Avoid "database is locked" error: don't open the two databases at the same time
+                try! dbQueue.inDatabase { db in
+                    try db.execute("BEGIN DEFERRED TRANSACTION")
+                    s1.signal()
+                    _ = s2.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 0)  // Zero balance
+                    s3.signal()
+                    _ = s4.wait(timeout: .distantFuture)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT SUM(value) AS balance FROM moves"), 0)  // Zero balance
+                    try db.execute("END")
+                }
+            }
+            let block2 = { () in
+                _ = s0.wait(timeout: .distantFuture) // Avoid "database is locked" error: don't open the two databases at the same time
+                let dbQueue = try! self.makeDatabaseQueue(filename: "test.sqlite")
+                try! dbQueue.inDatabase { db in
+                    _ = s1.wait(timeout: .distantFuture)
+                    try db.execute("BEGIN DEFERRED TRANSACTION")
+                    try db.execute("INSERT INTO moves VALUES (1)")
+                    s2.signal()
+                    _ = s3.wait(timeout: .distantFuture)
+                    try db.execute("INSERT INTO moves VALUES (-1)")
+                    s4.signal()
+                    try db.execute("END")
+                }
+            }
+            let blocks = [block1, block2]
+            DispatchQueue.concurrentPerform(iterations: blocks.count) { index in
+                blocks[index]()
+            }
+        }
+    }
+    
     func testWrappedReadWrite() {
         assertNoError {
             let dbPool = try makeDatabasePool()
@@ -343,13 +545,13 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             let block2 = { () in
                 try! dbPool.read { db in
                     _ = s1.wait(timeout: .distantFuture)
-                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 1)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 0)
                     s2.signal()
                     _ = s3.wait(timeout: .distantFuture)
-                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 1)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 0)
                     s4.signal()
                     _ = s5.wait(timeout: .distantFuture)
-                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 1)
+                    XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 0)
                 }
             }
             let blocks = [block1, block2]
@@ -422,7 +624,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
         }
     }
     
-    func testNonIsolatedReadMethodIsolationOfStatement() {
+    func testUnsafeReadMethodIsolationOfStatement() {
         assertNoError {
             let dbPool = try makeDatabasePool()
             try dbPool.write { db in
@@ -433,7 +635,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             }
             
             // Block 1                      Block 2
-            // dbPool.nonIsolatedRead {
+            // dbPool.unsafeRead {
             // SELECT * FROM items
             // step
             // >
@@ -447,7 +649,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             // }
             
             let block1 = { () in
-                try! dbPool.nonIsolatedRead { db in
+                try! dbPool.unsafeRead { db in
                     let cursor = try Row.fetchCursor(db, "SELECT * FROM items")
                     XCTAssertTrue(try cursor.next() != nil)
                     s1.signal()
@@ -475,7 +677,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
         }
     }
     
-    func testNonIsolatedReadMethodIsolationOfStatementWithCheckpoint() {
+    func testUnsafeReadMethodIsolationOfStatementWithCheckpoint() {
         assertNoError {
             let dbPool = try makeDatabasePool()
             try dbPool.write { db in
@@ -486,7 +688,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             }
             
             // Block 1                      Block 2
-            // dbPool.nonIsolatedRead {
+            // dbPool.unsafeRead {
             // SELECT * FROM items
             // step
             // >
@@ -500,7 +702,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             // }
             
             let block1 = { () in
-                try! dbPool.nonIsolatedRead { db in
+                try! dbPool.unsafeRead { db in
                     let cursor = try Row.fetchCursor(db, "SELECT * FROM items")
                     XCTAssertTrue(try cursor.next() != nil)
                     s1.signal()
@@ -528,7 +730,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
         }
     }
     
-    func testNonIsolatedReadMethodIsolationOfBlock() {
+    func testUnsafeReadMethodIsolationOfBlock() {
         assertNoError {
             let dbPool = try makeDatabasePool()
             try dbPool.write { db in
@@ -536,7 +738,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             }
             
             // Block 1                      Block 2
-            // dbPool.nonIsolatedRead {
+            // dbPool.unsafeRead {
             // SELECT COUNT(*) FROM items -> 0
             // >
             let s1 = DispatchSemaphore(value: 0)
@@ -548,7 +750,7 @@ class DatabasePoolConcurrencyTests: GRDBTestCase {
             // }
             
             let block1 = { () in
-                try! dbPool.nonIsolatedRead { db in
+                try! dbPool.unsafeRead { db in
                     XCTAssertEqual(try Int.fetchOne(db, "SELECT COUNT(*) FROM items")!, 0)
                     s1.signal()
                     _ = s2.wait(timeout: .distantFuture)